Add an option capture_stderr to control forked process stderr
If this is set, the subprocess's stderr stream will _NOT_ be redirected
to multiproc_log_device. This is normally not what you want for a
production app, but it's useful for my benchmarking efforts, so we may
as well commit this (but I didn't add a CLI option for it).